Refactor zone database
